# src/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Doormat",
    "description": "AI-first rental finder - autonomous property manager discovery and listing scoring",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/listings": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Get Listings",
        "description": "Return paginated, filterable listings.",
        "operationId": "get_listings_api_listings_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Price"
            }
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Price"
            }
          },
          {
            "name": "min_bedrooms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Bedrooms"
            }
          },
          {
            "name": "max_bedrooms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Bedrooms"
            }
          },
          {
            "name": "saved_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Saved Only"
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingResponse"
                  },
                  "title": "Response Get Listings Api Listings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/stream": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Stream Listings",
        "description": "SSE stream \u2014 push new listings as they arrive.",
        "operationId": "stream_listings_api_listings_stream_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/score": {
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Score Listings",
        "description": "Score persisted listings against a saved preference.",
        "operationId": "score_listings_api_listings_score_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreListingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreListingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/{listing_id}": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Get Listing",
        "description": "Return a single listing by ID.",
        "operationId": "get_listing_api_listings__listing_id__get",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Listing Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/{listing_id}/save": {
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Toggle Save Listing",
        "description": "Toggle the saved state of a listing.",
        "operationId": "toggle_save_listing_api_listings__listing_id__save_post",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Listing Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "property_manager_id": {
            "type": "string",
            "title": "Property Manager Id"
          },
          "preference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference Id"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "bedrooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bedrooms"
          },
          "bathrooms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bathrooms"
          },
          "sqft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqft"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "pets_policy": {
            "$ref": "#/components/schemas/PetsPolicy"
          },
          "amenities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Amenities"
          },
          "photos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photos"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "extraction_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Extraction Timestamp"
          },
          "extraction_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Model"
          },
          "tier1_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier1 Cost"
          },
          "tier2_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier2 Cost"
          },
          "validation_passed": {
            "type": "boolean",
            "title": "Validation Passed"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "score_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Explanation"
          },
          "saved": {
            "type": "boolean",
            "title": "Saved",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_manager_id",
          "address",
          "price",
          "url",
          "pets_policy",
          "amenities",
          "photos",
          "extraction_timestamp",
          "validation_passed"
        ],
        "title": "ListingResponse",
        "description": "Listing response model."
      },
...
      "PetsPolicy": {
        "type": "string",
        "enum": [
          "allowed_with_small_dog",
          "cats_only",
          "none_allowed",
          "unknown"
        ],
        "title": "PetsPolicy"
      },
...
      "ScoreListingsRequest": {
        "properties": {
          "preference_id": {
            "type": "string",
            "title": "Preference Id"
          },
          "listing_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Listing Ids"
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 50
          },
          "rescore": {
            "type": "boolean",
            "title": "Rescore",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "preference_id"
        ],
        "title": "ScoreListingsRequest",
        "description": "Run preference scoring for persisted listings."
      },
      "ScoreListingsResponse": {
        "properties": {
          "preference_id": {
            "type": "string",
            "title": "Preference Id"
          },
          "scored_count": {
            "type": "integer",
            "title": "Scored Count"
          },
          "listing_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Listing Ids"
          }
        },
        "type": "object",
        "required": [
          "preference_id",
          "scored_count",
          "listing_ids"
        ],
        "title": "ScoreListingsResponse",
        "description": "Summary of a listing scoring run."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}